Extract TopGamesForm to it's own component
